+++ common/autoware_universe_utils/test/src/geometry/test_geometry.cpp
@@ -1,4 +1,4 @@
-// Copyright 2020 Tier IV, Inc.
+// Copyright 2020-2024 Tier IV, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,6 +14,8 @@
 
 #include "autoware/universe_utils/geometry/boost_geometry.hpp"
 #include "autoware/universe_utils/geometry/geometry.hpp"
+#include "autoware/universe_utils/geometry/gjk_2d.hpp"
+#include "autoware/universe_utils/geometry/random_convex_polygon.hpp"
 #include "autoware/universe_utils/math/unit_conversion.hpp"
 #include "autoware/universe_utils/system/stop_watch.hpp"
 
@@ -24,9 +26,7 @@
 
 #include <gtest/gtest.h>
 
-#include <algorithm>
 #include <chrono>
-#include <random>
 #include <string>
 
 constexpr double epsilon = 1e-6;
@@ -1838,90 +1838,6 @@ TEST(geometry, intersect)
   }
 }
 
-/// @details taken from https://cglab.ca/~sander/misc/ConvexGeneration/convex.html
-// TODO(Maxime): move to test utils ?
-autoware::universe_utils::Polygon2d random_convex_polygon(const size_t vertices, const double max)
-{
-  using autoware::universe_utils::Point2d;
-  using autoware::universe_utils::Polygon2d;
-
-  struct VectorsWithMin
-  {
-    std::vector<double> vectors;
-    double min;
-  };
-
-  // static std::random_device r;
-  static std::default_random_engine e1(42);
-  static std::uniform_real_distribution<double> uniform_dist(-max, max);
-  static std::uniform_int_distribution random_bool(0, 1);
-  const auto prepare_coordinate_vectors = [&]() {
-    std::vector<double> v;
-    for (auto i = 0UL; i < vertices; ++i) {
-      v.push_back(uniform_dist(e1));
-    }
-    std::sort(v.begin(), v.end());
-    const auto min_v = v.front();
-    const auto max_v = v.back();
-    std::vector<double> v1;
-    v1.push_back(min_v);
-    std::vector<double> v2;
-    v2.push_back(min_v);
-    for (auto i = 1UL; i + 1 < v.size(); ++i) {
-      if (random_bool(e1) == 0) {
-        v1.push_back((v[i]));
-      } else {
-        v2.push_back((v[i]));
-      }
-    }
-    v1.push_back(max_v);
-    v2.push_back(max_v);
-    std::vector<double> diffs;
-    for (auto i = 0UL; i + 1 < v1.size(); ++i) {
-      diffs.push_back(v1[i + 1] - v1[i]);
-    }
-    for (auto i = 0UL; i + 1 < v2.size(); ++i) {
-      diffs.push_back(v2[i] - v2[i + 1]);
-    }
-    VectorsWithMin vectors;
-    vectors.vectors = diffs;
-    vectors.min = min_v;
-    return vectors;
-  };
-  auto xs = prepare_coordinate_vectors();
-  auto ys = prepare_coordinate_vectors();
-  std::shuffle(ys.vectors.begin(), ys.vectors.end(), e1);
-  autoware::universe_utils::LinearRing2d vectors;
-  for (auto i = 0UL; i < xs.vectors.size(); ++i) {
-    vectors.emplace_back(xs.vectors[i], ys.vectors[i]);
-  }
-  std::sort(vectors.begin(), vectors.end(), [](const Point2d & p1, const Point2d & p2) {
-    return std::atan2(p1.y(), p1.x()) < std::atan2(p2.y(), p2.x());
-  });
-  auto min_x = max;
-  auto min_y = max;
-  auto x = 0.0;
-  auto y = 0.0;
-  autoware::universe_utils::LinearRing2d points;
-  for (const auto & p : vectors) {
-    points.emplace_back(x, y);
-    x += p.x();
-    y += p.y();
-    min_x = std::min(p.x(), min_x);
-    min_y = std::min(p.y(), min_y);
-  }
-  const auto shift_x = min_x - xs.min;
-  const auto shift_y = min_y - ys.min;
-  for (auto & p : points) {
-    p.x() += shift_x;
-    p.y() += shift_y;
-  }
-  Polygon2d poly;
-  poly.outer() = points;
-  boost::geometry::correct(poly);
-  return poly;
-}
-
 TEST(geometry, intersectPolygon)
 {
   {  // 2 triangles with intersection
@@ -1992,9 +1908,9 @@ TEST(geometry, intersectPolygonRand)
   std::vector<autoware::universe_utils::Polygon2d> polygons;
   constexpr auto polygons_nb = 100;
   constexpr auto vertices = 4;
-  constexpr auto max_values = 10;
+  constexpr auto max_values = 100;
   for (auto i = 0; i < polygons_nb; ++i) {
-    polygons.push_back(random_convex_polygon(vertices, max_values));
+    polygons.push_back(autoware::universe_utils::random_convex_polygon(vertices, max_values));
   }
 
   autoware::universe_utils::StopWatch<std::chrono::nanoseconds, std::chrono::nanoseconds> sw;
@@ -2006,7 +1922,7 @@ TEST(geometry, intersectPolygonRand)
       const auto ground_truth = boost::geometry::intersects(polygons[i], polygons[j]);
       ground_truth_ns += sw.toc();
       sw.tic();
-      const auto gjk = autoware::universe_utils::intersect(polygons[i], polygons[j]);
+      const auto gjk = autoware::universe_utils::gjk::intersect(polygons[i], polygons[j]);
       gjk_ns += sw.toc();
       if (ground_truth != gjk) {
         std::cout << "Failed for the 2 polygons: ";
